+++ dubbo-samples-async/src/main/java/org/apache/dubbo/samples/async/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.async;
+package org.apache.dubbo.samples.async;
 
 import java.io.File;
 import java.lang.reflect.Method;
