+++ dubbo-samples-async/src/main/java/org/apache/dubbo/samples/async/api/AsyncService.java
@@ -17,17 +17,13 @@
  *
  */
 
-package com.alibaba.dubbo.samples.async;
+package org.apache.dubbo.samples.async.api;
 
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+/**
+ * AsyncService
+ */
+public interface AsyncService {
 
-public class AsyncProvider {
-
-    public static void main(String[] args) throws Exception {
-        new EmbeddedZooKeeper(2181, false).start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/async-provider.xml"});
-        context.start();
-        System.in.read();
-    }
+    String sayHello(String name);
 
 }
